requirements.txt
google-generativeai==0.3.2
flask==3.0.2
openai==0.28.0